Don't cast sh_size
ip:debloat

+++ src/ElfReader.cpp
@@ -117,9 +117,9 @@ void ElfReader::read_dynamic_symbols()
             dynstr_indx = i;
     }
     // dynamic table
-    int num_symbols = static_cast<int>(sections[dynsym_indx].sh_size / sizeof(Elf64_Sym));
+    uint64_t num_symbols = sections[dynsym_indx].sh_size / sizeof(Elf64_Sym);
     file.seekg(sections[dynsym_indx].sh_offset, ios::beg);
-    for(int i = 0; i < num_symbols; i++)
+    for(uint64_t i = 0; i < num_symbols; i++)
     {
         Elf64_Sym symbol;
         file.read((char*)(&symbol), sizeof(Elf64_Sym));
@@ -148,9 +148,9 @@ void ElfReader::read_symbols()
             strtab_indx = i;
     }
     // other symbols
-    int num_symbols = static_cast<int>(sections[symtab_indx].sh_size / sizeof(Elf64_Sym));
+    uint64_t num_symbols = sections[symtab_indx].sh_size / sizeof(Elf64_Sym);
     file.seekg(sections[symtab_indx].sh_offset, ios::beg);
-    for(int i = 0; i < num_symbols; i++)
+    for(uint64_t i = 0; i < num_symbols; i++)
     {
         Elf64_Sym symbol;
         file.read((char*)(&symbol), sizeof(Elf64_Sym));
@@ -175,9 +175,9 @@ void ElfReader::read_relocations()
     {
         if(sections[section_index].sh_type == SHT_RELA)
         {
-            int num_rela = static_cast<int>(sections[section_index].sh_size / sizeof(Elf64_Rela));
+            uint64_t num_rela = sections[section_index].sh_size / sizeof(Elf64_Rela));
             file.seekg(sections[section_index].sh_offset, ios::beg);
-            for(int i = 0; i < num_rela; i++)
+            for(uint64_t i = 0; i < num_rela; i++)
             {
                 Elf64_Rela relocation;
                 file.read((char*)(&relocation), sizeof(Elf64_Rela));
@@ -197,9 +197,9 @@ void ElfReader::read_dynamic_section()
     {
         if(sections[section_index].sh_type == SHT_DYNAMIC)
         {
-            int num_entries = static_cast<int>(sections[section_index].sh_size / sizeof(Elf64_Dyn));
+            uint64_t num_entries = sections[section_index].sh_size / sizeof(Elf64_Dyn);
             file.seekg(sections[section_index].sh_offset, ios::beg);
-            for(int i = 0; i < num_entries; i++)
+            for(uint64_t i = 0; i < num_entries; i++)
             {
                 Elf64_Dyn dynamic_entry;
                 file.read((char*)(&dynamic_entry), sizeof(Elf64_Dyn));
